Removed updateFrame; Added handling of new tracks
updateFrame() is not a method in the TrackData code, but the TrackLinker
calls it if the timepoint it is trying to update is between the last
frame and the first frame of the track.

This functionality is disabled temporarily until the issue is fixed
(see Issue #376).

Added some code to handle the case when the track array is empty but
assignToTrack is called. In this case, new tracks should just be added
and not be subject to the linking step (since there are no tracks to
link to anyway).